Functions

Exported functions

JuliaWorkspaces.add_file! — Function

add_file!(jw::JuliaWorkspace, file::TextFile)

Add a file to the workspace. If the file already exists, it will throw an error.

JuliaWorkspaces.remove_file! — Function

remove_file!(jw::JuliaWorkspace, uri::URI)

Remove a file from the workspace. If the file does not exist, it will throw an error.

JuliaWorkspaces.remove_all_children! — Function

remove_all_children!(jw::JuliaWorkspace, uri::URI)

Remove all children of a folder from the workspace.

JuliaWorkspaces.get_text_files — Function

get_text_files(jw::JuliaWorkspace)

Get all text files from the workspace.

Returns

A set of URIs.

JuliaWorkspaces.get_julia_files — Function

get_julia_files(jw::JuliaWorkspace)

Get all Julia files from the workspace.

Returns

A set of URIs.

JuliaWorkspaces.has_file — Function

has_file(jw, uri)

Check if a file exists in the workspace.

JuliaWorkspaces.get_text_file — Function

get_text_file(jw::JuliaWorkspace, uri::URI)

Get a text file from the workspace. If the file does not exist, it will throw an error.

Returns

A TextFile struct.

JuliaWorkspaces.get_julia_syntax_tree — Function

get_julia_syntax_tree(jw::JuliaWorkspace, uri::URI)

Get the syntax tree of a Julia file from the workspace.

Returns

The tuple (tree, diagnostics), where tree is the syntax tree and diagnostics is a vector of Diagnostic structs.

JuliaWorkspaces.get_toml_syntax_tree — Function

get_toml_syntax_tree(jw::JuliaWorkspace, uri::URI)

Get the syntax tree of a TOML file from the workspace.

JuliaWorkspaces.get_diagnostic — Function

get_diagnostic(jw::JuliaWorkspace, uri::URI)

Get the diagnostics of a file from the workspace.

Returns

A vector of Diagnostic structs.

JuliaWorkspaces.get_packages — Function

get_packages(jw::JuliaWorkspace)

Get all packages from the workspace.

Returns

A set of URIs.

JuliaWorkspaces.get_projects — Function

get_projects(jw::JuliaWorkspace)

Get all projects from the workspace.

Returns

A set of URIs.

JuliaWorkspaces.get_test_items — Function

get_test_items(jw::JuliaWorkspace, uri::URI)

Get the test items that belong to a given URI of a workspace.

Returns

an instance of the struct TestDetails

get_test_items(jw::JuliaWorkspace)

Get all test items of the workspace jw.

Returns

an instance of the struct TestDetails

JuliaWorkspaces.get_test_env — Function

get_test_env(jw::JuliaWorkspace, uri::URI)

Get the test environment that belongs to the given uri of the workspace jw.

Returns

an instance of the struct JuliaTestEnv

Private functions

JuliaWorkspaces.get_files — Function

get_files(jw::JuliaWorkspace)

Get all files from the workspace.

Returns

A set of URIs.

JuliaWorkspaces.get_diagnostics — Function

get_diagnostics(jw::JuliaWorkspace)

Get all diagnostics from the workspace.

Returns

A vector of Diagnostic structs.

JuliaWorkspaces.update_file! — Function

update_file!(jw::JuliaWorkspace, file::TextFile)

Update a file in the workspace. If the file does not exist, it will throw an error.

URI helper functions (submodule URIs2)

JuliaWorkspaces.URIs2.unescapeuri — Function

unescapeuri(str)

Percent-decode a string according to the URI escaping rules.

JuliaWorkspaces.URIs2.escapeuri — Function

escapeuri(x)

Apply URI percent-encoding to escape special characters in x.

JuliaWorkspaces.URIs2._bytes — Function

_bytes(s::String)

Get a Vector{UInt8}, a vector of bytes of a string.

JuliaWorkspaces.URIs2.escapepath — Function

escapepath(path)

Escape the path portion of a URI, given the string path containing embedded / characters which separate the path segments.