Types

Exported types

JuliaWorkspaces.TextFile — Type

struct TextFile

A text file, consisting of its URI and content.

uri::URI: The URI of the file.

content::SourceText: The content of the file as SourceText.

JuliaWorkspaces.SourceText — Type

struct SourceText

A source text, consisting of its content, line indices, and language ID.

content::String

line_indices::Vector{Int}

language_id::String

Private types

JuliaWorkspaces.Diagnostic — Type

struct Diagnostic

A diagnostic struct, consisting of range, severity, message, and source.

range::UnitRange{Int64}

severity::Symbol

message::String

source::String

JuliaWorkspaces.JuliaPackage — Type

struct JuliaPackage

Details of a Julia package.

project_file_uri::URI

name::String

uuid::UUID

content_hash::UInt

JuliaWorkspaces.JuliaTestEnv — Type

struct JuliaTestEnv

Details of a Julia test environment.

package_name::String

package_uri::Union{URI,Nothing}

project_uri::Union{URI,Nothing}

env_content_hash::Union{UInt,Nothing}

JuliaWorkspaces.JuliaProject — Type

struct JuliaProject

Details of a Julia project.

project_file_uri::URI

manifest_file_uri::URI

content_hash::UInt

deved_packages::Dict{String,JuliaProjectEntryDevedPackage}

regular_packages::Dict{String,JuliaProjectEntryRegularPackage}

stdlib_packages::Dict{String,JuliaProjectEntryStdlibPackage}

JuliaWorkspaces.JuliaProjectEntryDevedPackage — Type

struct JuliaProjectEntryDevedPackage

Details of a Julia project entry for a developed package.

name::String

uuid::UUID

uri::URI

version::String

JuliaWorkspaces.JuliaProjectEntryRegularPackage — Type

struct JuliaProjectEntryRegularPackage

Details of a Julia project entry for a regular package.

name::String

uuid::UUID

version::String

git_tree_sha1::String

JuliaWorkspaces.JuliaProjectEntryStdlibPackage — Type

struct JuliaProjectEntryStdlibPackage

Details of a Julia project entry for a standard library package.

name::String

uuid::UUID

version::Union{Nothing,String}

JuliaWorkspaces.JuliaWorkspace — Type

struct JuliaWorkspace

A Julia workspace, consisting of a Salsa runtime.

runtime::Salsa.Runtime

JuliaWorkspaces.NotebookFile — Type

struct NotebookFile

A notebook file, consisting of its URI and cells.

uri::URI: The URI of the file.

cells::Vector{SourceText}: The cells of the notebook as a vector of SourceText.

JuliaWorkspaces.TestSetupDetail — Type

struct TestSetupDetail

Details of a test setup.

uri::URI

name::Symbol

kind::Symbol

code::String

range::UnitRange{Int}

code_range::UnitRange{Int}

JuliaWorkspaces.TestDetails — Type

struct TestDetails

Details of a test.

testitems::Vector{TestItemDetail}

testsetups::Vector{TestSetupDetail}

testerrors::Vector{TestErrorDetail}

JuliaWorkspaces.TestItemDetail — Type

struct TestItemDetail

Details of a test item.

uri::URI

id::String

name::String

code::String

range::UnitRange{Int}

code_range::UnitRange{Int}

option_default_imports::Bool

option_tags::Vector{Symbol}

option_setup::Vector{Symbol}

JuliaWorkspaces.TestErrorDetail — Type

struct TestErrorDetail

Details of a test error.

uri::URI

id::String

name::Union{Nothing,String}

message::String

range::UnitRange{Int}

JuliaWorkspaces.URIs2.URI — Type

struct URI

Details of a Unified Resource Identifier.

scheme::Union{Nothing, String}

authority::Union{Nothing, String}

path::String

query::Union{Nothing, String}

fragment::Union{Nothing, String}